Fix build for aarch64
:Release Notes:

:Detailed Notes:

:Testing Performed:
Build Successful, APIs verified.

:QA Notes:

:Issues Addressed
[PLAT-43916] Integrate PMS into image

Reviewed-on: http://gpro.lgsvl.com/188623
Tested-by: Rohit Anand <rohit.anand@lge.com>

+++ src/PmsLunaInterface/LunaPmsRoot.cpp
@@ -101,7 +101,7 @@ LunaPmsRoot::LunaPmsRoot(PmsConfig* pConfig, LSHandle *pLsHandle) :
 void LunaPmsRoot::RegisterIpc(LSHandle *pLsHandle)
 {
     PMSLunaCategoryContext *pCxt = PMSLunaCategoryContext::Instance();
-    pCxt->AddLunaCategoryContext(kPmsMsgCategoryRoot, (unsigned int)this);
+    pCxt->AddLunaCategoryContext(kPmsMsgCategoryRoot, this);
 
     LSError lserror;
     LSErrorInit(&lserror);
